VVDN_Icon/requirement.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
numpy==1.24.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
pluggy==1.2.0
Pygments==2.16.1
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
robotframework==6.1.1
robotframework-datadriver==1.9.0
robotframework-pythonlibcore==4.2.0
robotframework-seleniumlibrary==6.1.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==4.0.0
wsproto==1.2.0
xlrd==2.0.1
